Move `http_responder.h` to public header
