+++ src/apps/external_executor/external_executor.cpp
@@ -6,6 +6,7 @@
 #include "ccf/crypto/verifier.h"
 #include "ccf/entity_id.h"
 #include "ccf/http_consts.h"
+#include "ccf/http_responder.h"
 #include "ccf/json_handler.h"
 #include "ccf/kv/map.h"
 #include "ccf/service/tables/nodes.h"
@@ -15,7 +16,6 @@
 #include "executor_registration.pb.h"
 #include "http/http2_session.h"
 #include "http/http_builder.h"
-#include "http/http_responder.h"
 #include "kv.pb.h"
 #include "node/endpoint_context_impl.h"
 #include "node/rpc/rpc_context_impl.h"
